+++ src/backends/qr-bin-lt/index.ts
@@ -16,6 +16,14 @@ class QrBinLtDecoder implements BackendDecoder<Uint8Array> {
     return this.decoder.progress;
   }
 
+  get totalBytes(): number | undefined {
+    return this.decoder.totalBytes;
+  }
+
+  get bytesReceived(): number {
+    return this.decoder.bytesReceived;
+  }
+
   getResult(): Uint8Array {
     return this.decoder.getResult();
   }